Restructure files into different folders under src
