+++ src/lib/rollbar.js
