+++ src/webhook/handlers/__tests__/askingReplyRequestReason.test.js
@@ -1,8 +1,8 @@
-jest.mock('../../gql');
+jest.mock('../../../lib/gql');
 
 import askingReplyRequestReason from '../askingReplyRequestReason';
 import * as apiResult from '../__fixtures__/askingReplyRequestReason';
-import gql from '../../gql';
+import gql from '../../../lib/gql';
 import { REASON_PREFIX } from '../utils';
 
 it('handles the case when prefix does not match', async () => {
