+++ src/webhook/handlers/askingReplyRequestReason.js
@@ -1,4 +1,4 @@
-import gql from '../gql';
+import gql from '../../lib/gql';
 import { getArticleURL, REASON_PREFIX, createArticleShareReply } from './utils';
 
 export default async function askingArticleSubmission(params) {
