+++ src/webhook/handlers/choosingReply.js
@@ -1,5 +1,5 @@
 import { t } from 'ttag';
-import gql from '../gql';
+import gql from '../../lib/gql';
 import {
   createPostbackAction,
   createReferenceWords,
@@ -9,7 +9,7 @@ import {
   getLIFFURL,
   DOWNVOTE_PREFIX,
 } from './utils';
-import ga from '../ga';
+import ga from '../../lib/ga';
 
 export default async function choosingReply(params) {
   let { data, state, event, issuedAt, userId, replies, isSkipUser } = params;
